requirements.txt
python-consul==0.7.2
docker==3.1.3
